requirements.txt
playwright>=1.40.0
openai>=1.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
Pillow>=10.0.0
schedule>=1.2.0
rich>=13.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
jinja2>=3.1.0
python-multipart>=0.0.6
